Added fr,sp,it language files to assets and build.py

+++ build.py
@@ -25,6 +25,12 @@ P2GZ_CUSTOM_ASSETS_COMPRESSED = [
 P2GZ_CUSTOM_ASSETS_UNCOMPRESSED = [
     os.path.join(P2GZ_ASSETS, 'files', 'opening.bnr'),
     os.path.join(P2GZ_ASSETS, 'sys', 'boot.bin'),
+    os.path.join(P2GZ_ASSETS, 'files', 'message', 'mesRes_fra.szs'),
+    os.path.join(P2GZ_ASSETS, 'files', 'message', 'mesRes_spa.szs'),
+    os.path.join(P2GZ_ASSETS, 'files', 'message', 'mesRes_ita.szs'),
+    os.path.join(P2GZ_ASSETS, 'files', 'new_screen', 'fra'),
+    os.path.join(P2GZ_ASSETS, 'files', 'new_screen', 'spa'),
+    os.path.join(P2GZ_ASSETS, 'files', 'new_screen', 'ita'),
 ]
 
 # argument parsing
@@ -95,7 +101,11 @@ for path in P2GZ_CUSTOM_ASSETS_UNCOMPRESSED:
     # Copy existing uncompressed file
     if os.path.exists(iso_path):
         print(f'Copying {path} to {iso_path}')
-        shutil.copy(path, iso_path)
+        # if it's a file or a directory
+        if os.path.isfile(path):
+            shutil.copy(path, iso_path) # copy individual file
+        else: # its a directory
+            shutil.copytree(path, iso_path, dirs_exist_ok=True) # copy whole directory
 
     # Add new directory + file
     else:
